# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Garden Maintenance System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "200":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Auth response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /care/generate:
    post:
      operationId: generateCareRoutine
      tags: [care]
      summary: Generate (or regenerate) a care routine for a garden plant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCareRoutineBody"
      responses:
        "200":
          description: Generated care routine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareRoutine"
...
  /admin/users/{userId}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Admin — delete a user
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOk"
  /admin/catalog:
    get:
      operationId: adminListCatalog
      tags: [admin]
      summary: Admin — list catalog plants
      responses:
        "200":
          description: Catalog plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogPlant"
    post:
      operationId: adminCreateCatalogPlant
      tags: [admin]
      summary: Admin — create a catalog plant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatalogPlantBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogPlant"
  /admin/catalog/{plantId}:
    patch:
      operationId: adminUpdateCatalogPlant
      tags: [admin]
      summary: Admin — update a catalog plant
      parameters:
        - in: path
          name: plantId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCatalogPlantBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogPlant"
    delete:
      operationId: adminDeleteCatalogPlant
      tags: [admin]
      summary: Admin — delete a catalog plant
      parameters:
        - in: path
          name: plantId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOk"
...
components:
  schemas:
...
    SimpleOk:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    User:
      type: object
      required: [id, email, name, region, role, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        region: { type: string }
        role: { type: string, enum: [user, admin] }
        avatarUrl: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...
    SignupBody:
      type: object
      required: [email, password, name, region]
      properties:
        email: { type: string }
        password: { type: string }
        name: { type: string }
        region: { type: string }
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
...
    CatalogPlant:
      type: object
      required: [id, commonName, scientificName, category, difficulty, imageUrl, summary]
      properties:
        id: { type: integer }
        commonName: { type: string }
        scientificName: { type: string }
        category: { type: string }
        difficulty: { type: string, enum: [easy, medium, hard] }
        imageUrl: { type: string }
        summary: { type: string }
        sunlight: { type: string }
        waterFrequencyDays: { type: integer }
        idealTemperatureC: { type: string }
        soilType: { type: string }
...
    CareRoutine:
      type: object
      required: [id, gardenPlantId, wateringFrequencyDays, wateringAmountMl, sunlight, soilNotes, fertilizerNotes, generatedAt, tasks]
      properties:
        id: { type: integer }
        gardenPlantId: { type: integer }
        wateringFrequencyDays: { type: integer }
        wateringAmountMl: { type: integer }
        sunlight: { type: string }
        soilNotes: { type: string }
        fertilizerNotes: { type: string }
        generatedAt: { type: string, format: date-time }
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/CareTask"
    CareTask:
      type: object
      required: [title, frequency, description]
      properties:
        title: { type: string }
        frequency: { type: string }
        description: { type: string }
    GenerateCareRoutineBody:
      type: object
      required: [gardenPlantId]
      properties:
        gardenPlantId: { type: integer }
...
    CreateCatalogPlantBody:
      type: object
      required: [commonName, scientificName, category, difficulty, imageUrl, summary, sunlight, waterFrequencyDays, idealTemperatureC, soilType, description]
      properties:
        commonName: { type: string }
        scientificName: { type: string }
        category: { type: string }
        difficulty: { type: string, enum: [easy, medium, hard] }
        imageUrl: { type: string }
        summary: { type: string }
        sunlight: { type: string }
        waterFrequencyDays: { type: integer }
        idealTemperatureC: { type: string }
        soilType: { type: string }
        description: { type: string }
    UpdateCatalogPlantBody:
      type: object
      properties:
        commonName: { type: string }
        scientificName: { type: string }
        category: { type: string }
        difficulty: { type: string, enum: [easy, medium, hard] }
        imageUrl: { type: string }
        summary: { type: string }
        sunlight: { type: string }
        waterFrequencyDays: { type: integer }
        idealTemperatureC: { type: string }
        soilType: { type: string }
        description: { type: string }
...