# lib/api-zod/src/generated/types/catalogPlant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Garden Maintenance System API
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogPlantDifficulty } from "./catalogPlantDifficulty";

export interface CatalogPlant {
  id: number;
  commonName: string;
  scientificName: string;
  category: string;
  difficulty: CatalogPlantDifficulty;
  imageUrl: string;
  summary: string;
  sunlight?: string;
  waterFrequencyDays?: number;
  idealTemperatureC?: string;
  soilType?: string;
}